requirements.txt
Flask==2.2.5
Pillow
opencv-python
transformers
torch
python-dotenv
